PodcastCreator_agent/requirements.txt
pydub
crewai
crewai-tools
elevenlabs
python-dotenv
pydantic